+++ .gitignore
@@ -104,3 +104,5 @@ venv.bak/
 .mypy_cache/
 
 .DS_Store
+
+tutorial/tutorial_outputs/*
